feat: add diagnostics command
Add a diagnostic command that will check the rate limit status of the currently configured GitHub token. For now that's all it does, but it could be easily built out to surface more information to the user.

The command is inferenced-based for the moment, which could lead to issues when LLM endpoint not available, but at the same rate 'intelligent feedback' can be helpful. We can provide alternatives in the future.

Note that this also includes a basic setup for using different prompts, which could become useful for supporting various workflows in the future.